:truck: move minesweeper api files
