+++ src/pages/games/mine-sweeper/apis/useCreateMineSweeperRecordMutation.ts
