+++ src/pages/games/mine-sweeper/apis/useModesQuery.ts
